Improve handling of persistent disks
This splits out the existing code for disk creation and formatting from
deploy_gcp.sh into a separate script where it can potentially be used
independently.

The new script does the same set of operations as before
but extends that slightly to allow for automatically copying a directory
of files from the local machine to the newly formatted disk through the
temporary VM it is attached to. The local and remote paths are set using
the new local_files_path/remote_files_path config parameters.

This also modifies deploy_gcp.sh to use a separate temporary VM for each
deployment to avoid issues aroud sharing the name for all of the
deployments (e.g. if a VM is created and then deleted by one deployment,
it might not be immediately possible to reuse that name for another VM).