Update core service configuration
The core service loads a file called search_api_config.json which
defines the URL and endpoints of the search API service. This commit
adds a "template" version of the original file which is used to generate
a new version of search_api_config.json during a deployment with the
domain correctly configured for the search API service as set in
deploy_gcp_config.