Implement endSession in DialogAgentManager
The endSession method in this class is called when the instance is
removed from the cache, but was previously empty. It now iterates over
the list of agents it owns and calls their new endSession methods.

This is currently only useful for the WizardAgent, which will unregister
any active Firestore listener objects.